tests/wpt/tests/permissions-policy/experimental-features/permissions-policy-header-host-wildcard.https.sub.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
  test(function() {
    var wildcard_origin = 'https://*:{{ports[https][0]}}';
    assert_array_equals(
      document.permissionsPolicy.getAllowlistForFeature('fullscreen'),
      [wildcard_origin]);
    assert_true(document.permissionsPolicy.allowsFeature('fullscreen'));
  }, 'Test wildcard host policy works as expected');
  </script>
</body>
